Progressive download for previews of images, audio, and video
- Add logic to `FilePreview` to skip pre-downloading the file for one of the
  above types, instead letting the rendered HTML tag handle download
- Update props for `NativePreview` to take a `string` value for `file` instead
  of `File`
- Update `NativePreview` component to pass `file` prop directly to HTML tags
